# swagger.yaml
swagger: "2.0"
info:
  description: |
    Backend API to support HotelHopper.cf \
    To see the standardized api responses see the models at the bottom\: 
    SuccessfulResponse for 200, BadRequestResponse for 400, and UnauthorizedResponse for 401 \
    More HTTP Status Codes may be added in the future and similar models will be made avialable.
  version: "1.0.0"
  title: "Hotel Hopper"
host: "staging-api.hotelhopper.cf"
...
schemes:
- "https"
paths:
...
  /reservations/:
    post:
      tags:
      - "reservations"
      summary: "Create a new reservation"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
      - name: body
        in: body
        schema:
          $ref: '#/definitions/CreateReservationRequest'
      responses:
        200:
          description: "OK"
          schema:
            type: object
            example:
              error: false
              message: "Your reservation has been successfully created!"
        400:
          description: "Bad Request"
          schema:
            $ref: '#/definitions/BadRequestResponse'
        401:
          description: "Unauthorized"
          schema:
            $ref: '#/definitions/UnauthorizedResponse'
    get:
      tags:
      - "reservations"
      summary: "Get all your reservations"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      responses:
        200:
          description: "OK"
          schema:
            $ref: '#/definitions/ReservationListResponse'
        400:
          description: "Bad Request"
          schema:
            $ref: '#/definitions/BadRequestResponse'
        401:
          description: "Unauthorized"
          schema:
            $ref: '#/definitions/UnauthorizedResponse'
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
...
  BadRequestResponse:
    type: object
    properties:
      error:
        type: boolean
      message:
        type: string
        description: "The first found validation error"
      validationErrors:
        type: object
        properties:
          <name of parameter>:
            type: array
            description: "Array of the problems with the parameter value"
            items:
              type: string
    required:
    - error
    - message
  UnauthorizedResponse:
    type: object
    properties:
      error:
        type: boolean
      message:
        type: string
        enum:
          - "No user found with JWT"
          - "Invalid JWT format"
          - "Expired JWT"
    required:
    - error
    - message
...
  ReservationListResponse:
    type: object
    properties:
      error:
        type: boolean
      data:
        type: array
        items:
          $ref: '#/definitions/ReservationDetail'
  CreateReservationRequest:
    type: object
    properties:
      hotelId:
        type: integer
      startDate:
        type: string
      endDate:
        type: string
      stripeToken:
        type: string
      rooms:
        type: array
        items:
          type: object
          properties:
            roomTypeId:
              type: string
            count:
              type: integer
  ReservationDetail:
    type: object
    properties:
      reservationId:
        type: integer
      hotelId:
        type: integer
      startDate:
        type: string
      endDate:
        type: string
      totalCost:
        type: integer
      status:
        type: string
      reservedRooms:
        type: array
        items:  
          type: object
          properties:
            count:
              type: integer
            roomType:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                persons:
                  type: integer
                beds:
                  type: integer
...